fix(installer): handle Squirrel hooks before app startup
